Fixed compile error.

+++ UIColorExtensions.swift
@@ -8,10 +8,10 @@
 import UIKit
 
 extension UIColor {
-    convenience init(hex: Int, alpha: Double = 1.0) {
-        let red = Double((hex & 0xFF0000) >> 16) / 255.0
-        let green = Double((hex & 0xFF00) >> 8) / 255.0
-        let blue = Double((hex & 0xFF)) / 255.0
+    convenience init(hex: Int, alpha: Float = 1.0) {
+        let red = Float((hex & 0xFF0000) >> 16) / 255.0
+        let green = Float((hex & 0xFF00) >> 8) / 255.0
+        let blue = Float((hex & 0xFF)) / 255.0
         self.init(red:red, green:green, blue:blue, alpha:alpha)
     }
 }
